// src/mentorado-ssi/enums/mss-indicador.enum.ts
// src/mentorado-ssi/enums/mss-indicador.enum.ts
export enum MssIndicador {
  // Pilar SSI (mantidos como indicadores, mas você pode removê-los se preferir)
  SSI_SETOR = 'SSI_SETOR',
  SSI_REDE = 'SSI_REDE',
  SSI_TOTAL = 'SSI_TOTAL',
  PILAR_MARCA = 'PILAR_MARCA',
  PILAR_PESSOAS_CERTAS = 'PILAR_PESSOAS_CERTAS',
  PILAR_INSIGHTS = 'PILAR_INSIGHTS',
  PILAR_RELACIONAMENTOS = 'PILAR_RELACIONAMENTOS',

  // Alcance/Perfil
  IMPRESSOES_PUBLICACAO = 'IMPRESSOES_PUBLICACAO',
  VISUALIZACOES_PERFIL = 'VISUALIZACOES_PERFIL',
  OCORRENCIAS_PESQUISA = 'OCORRENCIAS_PESQUISA',
  CARGOS_ENCONTRARAM_PERFIL = 'CARGOS_ENCONTRARAM_PERFIL',
  TAXA_RECRUTADORES = 'TAXA_RECRUTADORES',

  // Candidaturas
  CANDIDATURAS_SIMPLIFICADAS = 'CANDIDATURAS_SIMPLIFICADAS',
  CANDIDATURAS_VISUALIZADAS = 'CANDIDATURAS_VISUALIZADAS',
  CURRICULOS_BAIXADOS = 'CURRICULOS_BAIXADOS',
  CONTATOS_RH = 'CONTATOS_RH',

  // Conteúdo/Interações
  PUBLICACOES_SEMANA = 'PUBLICACOES_SEMANA',
  INTERACOES_COMENTARIOS = 'INTERACOES_COMENTARIOS',

  // Networking
  PEDIDOS_CONEXAO_HEADHUNTERS = 'PEDIDOS_CONEXAO_HEADHUNTERS',
  PEDIDOS_CONEXAO_DECISORES = 'PEDIDOS_CONEXAO_DECISORES',
  MENSAGENS_RECRUTADORES = 'MENSAGENS_RECRUTADORES',
  MENSAGENS_NETWORKING = 'MENSAGENS_NETWORKING',
  CAFES_AGENDADOS = 'CAFES_AGENDADOS',
  CAFES_TOMADOS = 'CAFES_TOMADOS',

  // Entrevistas / Ofertas
  ENTREVISTAS_REALIZADAS = 'ENTREVISTAS_REALIZADAS',
  ENTREVISTAS_FASE_FINAL = 'ENTREVISTAS_FASE_FINAL',
  CARTAS_OFERTA = 'CARTAS_OFERTA',
}